Crosswalk an effect estimate

Crosswalk an effect estimate

Usage

do_crosswalk(
  object,
  est_mean = NULL,
  est_se = NULL,
  est_ci = NULL,
  est_alpha = 0.05,
  alpha = 0.05,
  est_indep = NULL,
  est_outcome = NULL
)

Arguments

object

An object of class cogxwalkr or the result of est_cw_coef()

est_mean

Point estimate (beta) to be crosswalked to the alternative outcome measure

est_se

The standard error of est_mean

est_ci

The lower (1-alpha)% confidence interval of est_mean

est_alpha

The alpha level for the confidence interval (if est_se is provided) or the alpha level that will be used to back-calculate the standard error from est_ci. Defaults to 0.05.

alpha

The alpha level for the confidence interval of the crosswalked estimate. Defaults to 0.05.

est_indep

The independent variable to which est_mean applies

est_outcome

The outcome measure in the original study (e.g., "MOCA", "MMSE")